moving comment serializer to messaging module
